feat: Separate knowledge integration tests from unit tests (#520)
Extracts knowledge index integration tests into a dedicated test suite
to allow independent testing and better CI organization.

## Changes

**New Test Suite** (`tests/knowledge_integration.rs`):
- Moved 4 integration tests from `src/knowledge/index.rs`
- All tests require model downloads (marked `#[ignore]`)
- Tests: index creation, record/search, corrections, clear
- Feature-gated with `#[cfg(feature = "knowledge")]`

**Source Cleanup** (`src/knowledge/index.rs`):
- Removed inline integration tests
- Added comment pointing to new test location

**Documentation** (`tests/INTEGRATION_TESTS.md`):
- Comprehensive guide for integration test organization
- Feature flag alignment principles
- CI/CD integration examples
- Instructions for running knowledge tests

## Benefits

1. **Independent Testing**: Knowledge tests run separately from unit tests
2. **Feature Flag Alignment**: Tests match `knowledge` feature flag
3. **CI Flexibility**: Different test suites can run in separate CI jobs
4. **Better Organization**: Clear separation of concerns

## Running Tests

```bash
# Run knowledge integration tests (requires feature flag)
cargo test --test knowledge_integration --features knowledge -- --ignored
```

## Related
- Addresses #520
- Prepares for ChromaDB test suite (#519, #504)
