+++ src/knowledge/index.rs
@@ -358,75 +358,5 @@ pub struct KnowledgeStats {
     pub correction_count: usize,
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    #[tokio::test]
-    #[ignore = "requires model download"]
-    async fn test_index_create() {
-        let temp_dir = TempDir::new().unwrap();
-        let index = KnowledgeIndex::open(temp_dir.path()).await.unwrap();
-        let stats = index.stats().await.unwrap();
-        assert_eq!(stats.total_entries, 0);
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model download"]
-    async fn test_record_and_search() {
-        let temp_dir = TempDir::new().unwrap();
-        let index = KnowledgeIndex::open(temp_dir.path()).await.unwrap();
-
-        // Record a success
-        index
-            .record_success("list all files", "ls -la", Some("rust project"))
-            .await
-            .unwrap();
-
-        // Search for similar
-        let results = index.find_similar("show files", 5).await.unwrap();
-        assert!(!results.is_empty());
-        assert_eq!(results[0].command, "ls -la");
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model download"]
-    async fn test_record_correction() {
-        let temp_dir = TempDir::new().unwrap();
-        let index = KnowledgeIndex::open(temp_dir.path()).await.unwrap();
-
-        index
-            .record_correction(
-                "show disk usage",
-                "ls -lh",
-                "du -h -d 1",
-                Some("ls shows files not disk usage"),
-            )
-            .await
-            .unwrap();
-
-        let stats = index.stats().await.unwrap();
-        assert_eq!(stats.total_entries, 1);
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model download"]
-    async fn test_clear() {
-        let temp_dir = TempDir::new().unwrap();
-        let index = KnowledgeIndex::open(temp_dir.path()).await.unwrap();
-
-        index
-            .record_success("test", "echo test", None)
-            .await
-            .unwrap();
-
-        let stats = index.stats().await.unwrap();
-        assert_eq!(stats.total_entries, 1);
-
-        index.clear().await.unwrap();
-
-        let stats = index.stats().await.unwrap();
-        assert_eq!(stats.total_entries, 0);
-    }
-}
+// Tests for this module have been moved to tests/knowledge_integration.rs
+// Run with: cargo test --test knowledge_integration --features knowledge -- --ignored
